Add files via upload
An application that simulates the tossing of a coin. Each time
the user tosses the coin, the application uses a Random object to get a random integer in
the range of 0 through 1. If the random number is 0, it means the tails side of the coin
is up, and if the random number is 1, it means the heads side is up. The application
displays an image of a coin showing either heads or tails, depending on the value of the
random number.